+++ src/com/nutrons/recyclerush/commands/drivetrain/DriveHPIDCmd.java
@@ -1,16 +1,18 @@
-package com.nutrons.recyclerush.commands;
+package com.nutrons.recyclerush.commands.drivetrain;
 
 import com.nutrons.recyclerush.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
+ * 
+ * @author Camilo Gonzalez
  *
  */
-public class VisionCmd extends Command {
-
-    public VisionCmd() {
-        requires(Robot.camera);
+public class DriveHPIDCmd extends Command {
+	
+    public DriveHPIDCmd() {
+    	requires(Robot.dt);
     }
 
     // Called just before this Command runs the first time
@@ -19,7 +21,7 @@ public class VisionCmd extends Command {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	Robot.camera.getFeed();
+    	Robot.dt.drivePID(-Robot.oi.getJoystickX(), Robot.oi.getJoystickY(), Robot.oi.getJoystickSpin());
     }
 
     // Make this return true when this Command no longer needs to run execute()
@@ -29,10 +31,12 @@ public class VisionCmd extends Command {
 
     // Called once after isFinished returns true
     protected void end() {
+    	Robot.dt.stop();
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
+    	end();
     }
 }
